Allow to see editor hovers while debugging
Currently, when debugging, hovers provided by the LSP server are not displayed, because the are disabled at the start of debugging.
This change will disable these hovers only while showing a debug hover.
Thus during debugging, debug hovers have priority over editor hovers.

Fixes #15170
